Runner with some calculations

+++ Calculator/src/Data/Calculator.java
@@ -14,8 +14,8 @@ public class Calculator {
         return num1 * num2;
     }
 
-    public int divitionMethod(int num1, int num2) {
-        return num1 / num2;
+    public double divitionMethod(int num1, int num2) {
+        return (double) num1 / num2;
     }
 
 }
